+++ generators/component-hello-world-v2-ascii/templates/src/index.js
@@ -1,14 +1,16 @@
-const q         = require('q');
-const Component = require('component');
-const winston   = require('winston');
+const q             = require('q');
+const Component     = require('component');
+const winston       = require('winston');
 const winstonLogzio = require('winston-logzio');
+// const imageToAscii  = require("image-to-ascii");
+const imageToAscii  = require("asciify-image");
 
 
 // Component implementation.
 //
 // Constructor, run, shutdown and configure methods are invoked by Kumori PaaS,
 // managing instance lifecycle.
-class DataStorage extends Component {
+class AsciiConverter extends Component {
 
 
   // Constructor.
@@ -79,17 +81,17 @@ class DataStorage extends Component {
       logzioWinstonTransport.flush(() => process.exit(1));
     }).bind(this));
 
-    this.logger.info(`DataStorage.constructor role=${this.role} iid=${this.iid}`);
+    this.logger.info(`AsciiConverter.constructor role=${this.role} iid=${this.iid}`);
 
     // This component uses a Reply channel, through which the requests coming
     // from FE role arrive.
-    const dataapiChannel = this.offerings['dataapi'];
-    if (dataapiChannel == null) {
-      throw new Error('Dataclient channel not found');
+    const asciiapiChannel = this.offerings['asciiapi'];
+    if (asciiapiChannel == null) {
+      throw new Error('Asciiapi channel not found');
     }
 
     // We assign a custom function that handles requests.
-    dataapiChannel.handleRequest = this._handleDataapiRequest.bind(this);
+    asciiapiChannel.handleRequest = this._handleRequest.bind(this);
 
     this._storage = {};
   }
@@ -98,7 +100,7 @@ class DataStorage extends Component {
   // Method invoked by Kumori PaaS to start instance execution.
   // Not special to do in this example.
   run() {
-    this.logger.info('DataStorage.run');
+    this.logger.info('AsciiConverter.run');
     super.run();
   }
 
@@ -109,7 +111,7 @@ class DataStorage extends Component {
   // shutdown, it will be killed, so be sure to fill this method.
   // Not special to do in this example, as we store data in memory.
   shutdown() {
-    this.logger.info('DataStorage.shutdown');
+    this.logger.info('AsciiConverter.shutdown');
     super.shutdown();
   }
 
@@ -117,7 +119,7 @@ class DataStorage extends Component {
   // Method invoked by Kumori PaaS to modify instance configuration.
   // Not really used in this example.
   reconfig(parameters) {
-    this.logger.info('DataStorage.reconfig');
+    this.logger.info('AsciiConverter.reconfig');
     super.reconfig(parameters);
   }
 
@@ -126,63 +128,30 @@ class DataStorage extends Component {
   //    handleRequest = (message, dynamicChannels) -> ...
   //
   // - Message: array of segments (buffers) received in request.
-  //    In this example we have a single segment, that is a
-  //    serialized JSON with the following attributes:
-  //     - a string with the requested operation.
-  //     - a key
-  //     - a value (only if operation is 'write')
+  //    In this example we have a single segment containing the image data.
   // - dynamicChannels: array of dynamic channels.
   //    Not used in this example.
   //
   // It returns a promise that is either:
-  // - Resolved with an array of buffers.
+  // - Resolved with the image converted to asciiart.
   // or
   // - Rejected with an error.
-  _handleDataapiRequest(message, dynamicChannels) {
-    this.logger.info('DataStorage._handleRequest');
+  _handleRequest(message, dynamicChannels) {
+    this.logger.info('AsciiConverter._handleRequest');
 
     return q.Promise((resolve, reject) => {
+
       try {
-        message = JSON.parse(message[0].toString());
-
-        let result;
-
-        if (message.operation == null) {
-          throw new Error('Undefined operation');
-        }
-        if (message.key == null) {
-          throw new Error('Undefined key');
-        }
-
-        switch (message.operation) {
-          case 'read':
-            let value = this._storage[message.key];
-            if (value == null) {
-              value = 'not found';
-            }
-            result = {
-              value,
-              instance: this.iid
-            };
-            break;
-
-          case 'write':
-            if (message.value == null) {
-              throw new Error('Undefined value');
-            }
-            this._storage[message.key] = message.value;
-            result = {
-              instance: this.iid
-            };
-            break;
-
-          default:
-            throw new Error(`Invalid operation (${message.operation})`);
-        }
-        resolve([JSON.stringify(result)]);
+        imageToAscii(message[0], {color: false, width: 200, height: 100, fit: 'box'}, (err, converted) => {
+          if (err) {
+            this.logger.warn(`AsciiConverter._handleRequest ${err.message || err}`);
+            reject(err)
+          } else {
+            resolve([[new Buffer(converted)]])
+          }
+        })
       } catch (err) {
-        this.logger.warn(`DataStorage._handleRequest ${err.message}`);
-
+        this.logger.warn(`AsciiConverter._handleRequest ${err.message}`);
         reject(err);
       }
     });
@@ -190,4 +159,4 @@ class DataStorage extends Component {
 }
 
 
-module.exports = DataStorage;
+module.exports = AsciiConverter;