+++ generators/component-hello-world-v2-fe/templates/src/index.js
@@ -9,6 +9,7 @@ const winstonLogzio = require('winston-logzio');
 //
 // Constructor, run, shutdown and configure methods are invoked by Kumori PaaS,
 // managing instance lifecycle.
+
 class Fe extends Component {
 
 
@@ -40,6 +41,7 @@ class Fe extends Component {
   //    Keys are channel names, values are channel objects.
   // - offerings: dictionary of the channels offered by the component, through
   //    which you can answer requests from other roles or services.
+
   constructor(runtime, role, iid, incnum, localData, resources,
                 parameters, dependencies, offerings ) {
 
@@ -90,22 +92,22 @@ class Fe extends Component {
     }
 
     // This component also uses a Request channel, through which the requests
-    // to the other role (dataStorage) within the service are issued.
-    const dataclientChannel = this.dependencies['dataclient'];
-    if (dataclientChannel == null){
-      throw new Error('Dataclient channel not found');
+    // to the other role (ascii converter) within the service are issued.
+    const asciiChannel = this.dependencies['asciiclient'];
+    if (asciiChannel == null){
+      throw new Error('Asciiclient channel not found');
     }
 
     // RestApi is the object that will in truth attend HTTP request.
     // It internally uses Express module (see restapi.js for details).
     // It is given the channel through which the HTTP request arrive.
-    // It is given the channel to issue requests to dataStorage role.
-    this.restapi = new RestApi(entrypointChannel, dataclientChannel
-      , this.logger);
+    // It is given the channel to issue requests to ascii converter role.
+    this.restapi = new RestApi(this.iid, entrypointChannel, asciiChannel, this.logger);
   }
 
 
   // Method invoked by Kumori PaaS to start instance execution.
+
   run() {
     this.logger.info('Fe.run');
     super.run();
@@ -128,6 +130,7 @@ class Fe extends Component {
   // shutdown. Instance should take necessary actions in this situation,
   // persisting its state if needed. If the instance doesn't gracefully
   // shutdown, it will be killed, so be sure to fill this method.
+
   shutdown() {
     this.logger.info('Fe.shutdown');
     this.restapi.stop()
@@ -140,6 +143,7 @@ class Fe extends Component {
 
   // Method invoked by Kumori PaaS to modify instance configuration.
   // Not really used in this example.
+
   reconfig(parameters) {
     this.logger.info('Fe.reconfig');