+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## v1.0.2
 
 Minor and decorative changes to some components code.
-Removed a duplicated _send method in `component-hello-world-v2-fe/templates/src/restapi.js` generator.
+Removed a duplicated `_send` method in `component-hello-world-v2-fe/templates/src/restapi.js` generator.
 Removed deprecated `.kumori` folder in generators templates.
+Changed the static webpage in hello-world component.
 
 ## v1.0.1