+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Minor and decorative changes to some components code.
 Removed a duplicated `_send` method in `component-hello-world-v2-fe/templates/src/restapi.js` generator.
 Removed deprecated `.kumori` folder in generators templates.
 Changed the static webpage in hello-world component.
+Changed the hello world v2 project to another one more usable.
 
 ## v1.0.1